requirements.txt
python-box
torch
matplotlib
pyyaml
Pillow
imageio